Remove `@media` query for dark color scheme
This change basically forces the dark theme to always be active. This commit
doesn't unindent the styles that were previously within the `@media` query, so
that I could have a cleaner diff, which will make it easier for me to make the
same change when I want to update the repo.

The branch name is suffixed with the short SHA of the commit from which it was
branched - so that my update flow could be as easy as fast-forwarding master,
checking out a new branch (with a new name including the new SHA), and repeat
the changes from this branch - which I think will be much easier in the long run
than trying to keep a single branch in sync via rebasing or merging, and wasting
time resolving conflicts.